auth

+++ app/Http/Controllers/ConflictController.php
@@ -4,17 +4,26 @@ namespace App\Http\Controllers;
 
 use App\Conflict;
 use App\Http\Requests\ConflictRequest;
+use Illuminate\Support\Facades\Auth;
 
 class ConflictController extends Controller
 {
     public function index()
     {
-        return Conflict::get();
+        return Conflict::with('user')->get();
     }
 
     public function store(ConflictRequest $request)
     {
-        return Conflict::create($request->validated())->toArray();
+        $this->authorize('create', Conflict::class);
+
+        $data = $request->validated();
+
+        $data['user_id'] = object_get(Auth::user(), 'id');
+
+        $conflict = Conflict::create($data);
+
+        return $conflict->fresh('user')->toArray();
     }
 
     public function show(Conflict $conflict)
@@ -23,19 +32,23 @@ class ConflictController extends Controller
 
         $conflict->save();
 
-        return $conflict;
+        return $conflict->fresh('user');
     }
 
     public function update(ConflictRequest $request, Conflict $conflict)
     {
+        $this->authorize('update', $conflict);
+
         $conflict->update($request->validated());
 
-        return $conflict->toArray();
+        return $conflict->fresh('user')->toArray();
     }
 
     public function destroy(Conflict $conflict)
     {
-        $conflict ->delete();
+        $this->authorize('delete', $conflict);
+
+        $conflict->delete();
 
         return $conflict->id;
     }
